+++ src/platform_impl/linux/mod.rs
@@ -22,7 +22,10 @@ use raw_window_handle::RawWindowHandle;
 #[cfg(feature = "x11")]
 pub use self::x11::XNotSupported;
 #[cfg(feature = "x11")]
-use self::x11::{ffi::XVisualInfo, util::WindowType as XWindowType, XConnection, XError};
+use self::x11::{
+    ffi::XVisualInfo, util::WindowStrut as XWindowStrut, util::WindowType as XWindowType,
+    XConnection, XError,
+};
 use crate::{
     dpi::{PhysicalPosition, PhysicalSize, Position, Size},
     error::{ExternalError, NotSupportedError, OsError as RootOsError},
@@ -67,6 +70,8 @@ pub struct PlatformSpecificWindowBuilderAttributes {
     pub x11_window_types: Vec<XWindowType>,
     #[cfg(feature = "x11")]
     pub gtk_theme_variant: Option<String>,
+    #[cfg(feature = "x11")]
+    pub x11_window_struts: Vec<XWindowStrut>,
     #[cfg(feature = "wayland")]
     pub app_id: Option<String>,
 }
@@ -90,6 +95,8 @@ impl Default for PlatformSpecificWindowBuilderAttributes {
             x11_window_types: vec![XWindowType::Normal],
             #[cfg(feature = "x11")]
             gtk_theme_variant: None,
+            #[cfg(feature = "x11")]
+            x11_window_struts: Vec::new(),
             #[cfg(feature = "wayland")]
             app_id: None,
         }